requirements.txt
arxiv>=2.1.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dateutil>=2.8.2
google-genai>=0.1.0
